requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# UI
streamlit>=1.28.0

# Configuration
pyyaml>=6.0
jsonschema>=4.17.0

# Utilities
python-dateutil>=2.8.2
